+ Basic IndoorScene Start

+++ Box2D_SFML1/Box2D_SFML1/ICubemon.cpp
@@ -1,6 +1,6 @@
 #include "ICubemon.h"
 
-int ICubemon::TakeDamage(float _amount)
+int ICubemon::TakeDamage(int _amount)
 {
 	if (m_DamageTimer.getElapsedTime().asSeconds() >= 0.1f)
 	{
@@ -27,7 +27,7 @@ int ICubemon::TakeDamage(float _amount)
 	}
 }
 
-int ICubemon::Heal(float _amount)
+int ICubemon::Heal(int _amount)
 {
 	if (m_DamageTimer.getElapsedTime().asSeconds() >= 0.1f)
 	{
